add conversation panel
